Rework directory and code structure
We now consider the system being built as 'toplevel group', and put the
output in a directory named after that group. Then each component goes
into its own subdirectory, along with any shell scripts associated.

I think this is tidier, and it also allows the component builds to be
wrapped with some kind of prebuilt artifact caching check.